Fix ByteList512KiB limit to be a plain int
The test-only ByteList512KiB declared LIMIT as Uint64(512 * 1024), but
BaseByteList compares the payload length against LIMIT with a plain int
on the left-hand side. Uint64 rejects mixed-type comparisons, so the
reflected __lt__ raised a TypeError and every ByteList filler failed.
Declare LIMIT as a plain int, matching the other limits in this file.

+++ tests/fillers/ssz/test_basic_types.py
@@ -37,7 +37,7 @@ class Bytes64(BaseBytes):
 
 
 class ByteList512KiB(BaseByteList):
-    LIMIT = Uint64(512 * 1024)
+    LIMIT = 512 * 1024
 
 
 class SampleBitvector8(BaseBitvector):
